Rename mini_divider.jsx -> miniDivider.jsx
