+++ src/divider/stories/index.jsx
@@ -1,6 +1,6 @@
 import { storiesOf } from '@storybook/react'
 
-import miniDivider from './mini_divider'
+import miniDivider from './miniDivider'
 
 storiesOf('Divider', module)
   .add('MiniDivider', miniDivider)
