+++ src/divider/stories/miniDivider.jsx
